Break out the ESPHome Bluetooth scanner connection logic into bleak-esphome
